+++ react/AffiliateValidator.tsx
@@ -6,6 +6,8 @@ import { useOrderForm } from 'vtex.order-manager/OrderForm'
 import IS_AFFILIATE_VALID_QUERY from './graphql/isAffiliateValid.graphql'
 import SET_ON_ORDER_FORM_MUTATION from './graphql/setAffiliateOnOrderForm.graphql'
 
+const DEFAULT_ORDER_FORM_ID = 'default-order-form'
+
 export type IsAffiliateValidQueryResult = {
   isAffiliateValid: boolean
 }
@@ -26,26 +28,28 @@ const AffiliateValidator: FC<Props> = ({ Invalid, Valid }) => {
     return splitPathname && splitPathname[splitPathname.length - 1]
   }, [])
 
-  const [setAffiliateOnOrderForm] = useMutation(SET_ON_ORDER_FORM_MUTATION)
+  const [setAffiliateOnOrderForm, { called: mutationHasBeenCalled }] =
+    useMutation(SET_ON_ORDER_FORM_MUTATION)
 
   const { data, error } = useQuery<IsAffiliateValidQueryResult>(
     IS_AFFILIATE_VALID_QUERY,
     {
       variables: { slug },
       skip: !slug,
-      onCompleted: () => {
-        if (data?.isAffiliateValid) {
-          setAffiliateOnOrderForm({
-            variables: { orderFormId, affiliateId: slug },
-          })
-        }
-      },
     }
   )
 
   const isAffiliateValid = data?.isAffiliateValid
 
-  if (isAffiliateValid) return <Valid />
+  if (isAffiliateValid) {
+    if (!mutationHasBeenCalled && orderFormId !== DEFAULT_ORDER_FORM_ID) {
+      setAffiliateOnOrderForm({
+        variables: { orderFormId, affiliateId: slug },
+      })
+    }
+
+    return <Valid />
+  }
 
   if (isAffiliateValid === false || error) return <Invalid />